+++ src/routes/thread/comment/pin.ts
@@ -63,7 +63,7 @@ export default function (
 
             if (!comment) return res.code(404).send({ error: "Comment not found." });
             if (comment.removed)
-                return res.code(410).send({ error: "Comment has been removed." });
+                return res.code(410).send({ error: "Comment removed." });
 
             await threadCl.updateOne({ id: threadId }, { $set: { pin: comment } });